feat: queue - refactor theme rendering
